requirements.txt
# Existing dependencies
youtube-transcript-api==0.6.2
yt-dlp==2024.12.6
sqlite-utils==3.37
python-dotenv==1.0.1
requests==2.32.3

# Course Builder dependencies
chromadb>=0.4.0
anthropic>=0.39.0
openai>=1.0.0
tiktoken>=0.5.0
pydantic>=2.0.0
pyyaml>=6.0
